% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_calidad_materiales.R
\name{calcular_calidad_materiales}
\alias{calcular_calidad_materiales}
\title{Calidad de los materiales}
\usage{
calcular_calidad_materiales(x)
}
\arguments{
\item{x}{dataset de encuesta de hogares armonizada}
}
\value{
dataset con el indicador 'cm' generado
}
\description{
A partir de esta función se genera la variable 'cm':
identifica la calidad de los materiales de la vivienda donde reside el hogar
Si los materiales son suficiente, entonces 'cm' toma valor 2.
Si los materiales son recuperables, entonces 'cm' toma valor 1.
Si los materiales son irrecuparables, entonces 'cm' toma valor 0.
}
\details{
En caso de que no haya datos suficientes, 'cm' no toma ningún valor, es decir es NA.
}
