% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_deficit_vivienda.R
\name{calcular_deficit_vivienda}
\alias{calcular_deficit_vivienda}
\title{Indicador de deficit de la vivienda}
\usage{
calcular_deficit_vivienda(x)
}
\arguments{
\item{x}{dataset de encuesta de hogares armonizada}
}
\value{
dataset con el indicador 'deficit' generado
}
\description{
Esta funcion calcula para cada hogar si habita en una vivienda con:
deficit cuantitativo
deficit cualitativo
sin deficit
}
\details{
El hogar reside en una vivienda con deficit cuantitativo si:
La calidad de los materiales es irrecuperable
El hacinamiento es critico
El tipo de vivienda es inadecuado
No cuenta con baño exclusivo para el hogar

Es suficiente cumplir con por lo menos un criterio

El hogar reside en una vivienda con déficit cuantitativo si:
La calidad de los materiales es recuperable
El hacinamiento es mitigable
No cuenta con abastecimiento de agua
No cuenta con saneamiento - alcantarillado / cloaca
No cuenta con electricidad

Es suficiente cumplir con por lo menos un criterio

La funcion genera la variable 'deficit':
Si el hogar cuenta con déficit cuantitativo, entonces 'deficit' toma valor 2.
Si el hogar cuenta con déficit cualitativo, entonces 'deficit' toma valor 1.
Si el hogar no tiene deficit, entonces 'deficit' toma valor 0.

En caso de que no haya datos suficientes, 'deficit' no toma ningún valor, es decir es NA.
}
