% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_disponibilidad_de_agua.R
\name{calcular_disponibilidad_de_agua}
\alias{calcular_disponibilidad_de_agua}
\title{Disponibilidad de agua}
\usage{
calcular_disponibilidad_de_agua(x)
}
\arguments{
\item{x}{dataset de encuesta de hogares armonizada}
}
\value{
dataset con el indicador 'ag' generado
}
\description{
A partir de esta funcion se genera la variable 'ag':
identifica la disponibilidad de agua de la vivienda donde reside el hogar
}
\details{
Si la provisión de agua es optima, entonces 'ag' toma valor 1.
Si la provisión de agua es deficiente, entonces 'ag' toma valor 0.

En caso de que no haya datos suficientes, 'ag' no toma ningún valor, es decir es NA.

Esta variable discrimina si la vivienda se ubica en zona rural o urbana.
}
