% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_disponibilidad_bano.R
\name{calcular_disponibilidad_de_bano}
\alias{calcular_disponibilidad_de_bano}
\title{Disponibilidad de baño}
\usage{
calcular_disponibilidad_de_bano(x)
}
\arguments{
\item{x}{dataset de encuesta de hogares armonizada}
}
\value{
dataset con el indicador 'db' generado
}
\description{
A partir de esta funcion se genera la variable 'db':
identifica si el hogar cuenta con un baño exclusivo
}
\details{
Si el hogar dispone de un baño exclusivo, entonces 'db' toma valor 1.
Si el hogar no dispone de un baño exclusivo, entonces 'db' toma valor 0.

En caso de que no haya datos suficientes, 'db' no toma ningún valor, es decir es NA.
}
