% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_electricidad.R
\name{calcular_electricidad}
\alias{calcular_electricidad}
\title{Acceso a electricidad}
\usage{
calcular_electricidad(x)
}
\arguments{
\item{x}{dataset de encuesta de hogares armonizada}
}
\value{
dataset con el indicador 'el' generado
}
\description{
A partir de esta funcion se genera la variable 'el':
identifica el acceso a electricidad de la vivienda donde reside el hogar
}
\details{
Si el acceso a la electricidad es optimo, entonces 'el' toma valor 1.
Si el acceso a la electricidad es deficiente, entonces 'el' toma valor 0.

En caso de que no haya datos suficientes, 'el' no toma ningún valor, es decir es NA.
}
