% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_hacinamiento.R
\name{calcular_hacinamiento}
\alias{calcular_hacinamiento}
\title{Hacinamiento}
\usage{
calcular_hacinamiento(x)
}
\arguments{
\item{x}{dataset de encuesta de hogares armonizada}
}
\value{
dataset con el indicador 'ha' generado
}
\description{
A partir de esta función se genera la variable 'ha':
identifica el hacinamiento del hogar a partir de la cantidad de personas por dormitorio
}
\details{
Si el hacinamiento es crítico, entonces 'ha' toma valor 2.
Si el hacinamiento es recuperable, entonces 'ha' toma valor 1.
Si el hogar no vive en hacinamiento, entonces 'ha' toma valor 0.

En caso de que no haya datos suficientes, 'ha' no toma ningún valor, es decir es NA.
}
