% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_saneamiento.R
\name{calcular_saneamiento}
\alias{calcular_saneamiento}
\title{Acceso a saneamiento}
\usage{
calcular_saneamiento(x)
}
\arguments{
\item{x}{dataset de encuesta de hogares armonizada}
}
\value{
dataset con el indicador 'cl' generado
}
\description{
A partir de esta funcion se genera la variable 'cl':
identifica el acceso a saneamiento  - alcantarillado / cloaca - de la vivienda donde reside el hogar
}
\details{
Si el saneamiento es optimo, entonces 'cl' toma valor 1.
Si el sanemmiento es deficiente, entonces 'cl' toma valor 0.

En caso de que no haya datos suficientes, 'cl' no toma ningún valor, es decir es NA.

Esta variable discrimina si la vivienda se ubica en zona rural o urbana.
}
