% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corregir_encoding.R
\name{corregir_encoding}
\alias{corregir_encoding}
\title{Correccion manual de encoding}
\usage{
corregir_encoding(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector con los caracteres especiales del castellano corregidos
}
\description{
Permite corregir manualmente los caracteres especiales en las encuestas de hogares 'Harmonized Surveys'
}
\examples{
corregir_encoding(c("Asunción", "Saã Paulo", "Tucumán"))

}
