% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_eh.R
\name{descargar_eh}
\alias{descargar_eh}
\title{Descarga de dataset desde un AWS bucket}
\usage{
descargar_eh(path)
}
\arguments{
\item{path}{ruta a la encuesta de hogares armonizada}
}
\value{
dataset raw descargado desde aws como tabla
}
\description{
El dataset en AWS debe ser un archivo dta
}
