% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determinar_variables_disponibles.R
\name{determinar_variables_disponibles}
\alias{determinar_variables_disponibles}
\title{Preprocesamiento de las encuestas de hogares}
\usage{
determinar_variables_disponibles(x)
}
\arguments{
\item{x}{dataframe de encuesta de hogares armonizada}
}
\value{
un número entre 1 y 3 que determina las variables a utilizar para las métricas de calidad
}
\description{
Verifica las variables disponibles para el cálculo de métricas de calidad en la encuesta harmonizada
}
