% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encuesta_hogar_arg_2015.R
\docType{data}
\name{encuesta_hogar_arg_2015}
\alias{encuesta_hogar_arg_2015}
\title{Dataset ejemplo de Harmonized Survey - encuesta armonizada - de SCL DATA.}
\format{
A tibble with 120173 rows and 419 variables:
}
\source{
SCL DATA
}
\usage{
encuesta_hogar_arg_2015
}
\description{
Se trata de la encuesta EPH de Argentina del año 2015. Incluye las variables originales y las variables armonizadas.
}
\keyword{datasets}
