% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimar_indicador_deficit.R
\name{estimar_indicador_deficit}
\alias{estimar_indicador_deficit}
\title{Calculo de Indicador de Deficit habitacional}
\usage{
estimar_indicador_deficit(path)
}
\arguments{
\item{path}{ruta a la encuesta de hogares armonizada}
}
\value{
dataset con el indicador para esa encuesta de hogares
}
\description{
Calculo de Indicador de Deficit habitacional
}
