% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_ruta_eh_scldata.R
\name{generar_ruta_eh_scldata}
\alias{generar_ruta_eh_scldata}
\title{Generador de rutas dentro de SCLDATA}
\usage{
generar_ruta_eh_scldata(anio = NA, pais = "GRAL")
}
\arguments{
\item{anio}{define el anio, por defecto: todos los anios}

\item{pais}{define el pais, por defecto: todos los paises}
}
\value{
lista con las rutas para descargar los dataset
}
\description{
La funcion genera una lista con las rutas de scldata para descargar las encuestas armonizadas de hogares.
Se puede configurar para un anio o un pais en particular
}
