% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocesar_eh_bid.R
\name{preprocesar_eh_bid}
\alias{preprocesar_eh_bid}
\title{Preprocesamiento de Encuestas de Hogares}
\usage{
preprocesar_eh_bid(data)
}
\arguments{
\item{data}{dataset de encuesta de hogares armonizada}
}
\value{
dataset el indicador de deficit generado, incluye subindicadores
}
\description{
La funcion toma como input una encuesta armonizada de SCLDATA y devuelve como output un dataset simplificado.
Además calcula el indicador de deficit para cada hogar de la encuesta.
}
