% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scldata_seleccionar_ruta.R
\name{scldata_seleccionar_ruta}
\alias{scldata_seleccionar_ruta}
\title{Generador de rutas dentro de SCLDATA}
\usage{
scldata_seleccionar_ruta(x, inicio = NA, final = inicio, isoalpha3 = "GRAL")
}
\arguments{
\item{inicio}{define el anio inicial, por defecto el valor es NA}

\item{final}{define el anio final, por defecto es el mismo valor que inicio}

\item{pais}{define el pais, por defecto es 'GRAL' es decir todos los paises}
}
\value{
lista con las rutas para descargar los dataset
}
\description{
La funcion genera una lista con las rutas de scldata para descargar las encuestas armonizadas de hogares.
Se puede configurar para un anio, un lapso de anios o un pais en particular
}
