% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seleccionar_variables.R
\name{seleccionar_variables}
\alias{seleccionar_variables}
\title{Preprocesamiento de encuestas de hogares}
\usage{
seleccionar_variables(x)
}
\arguments{
\item{x}{dataframe de encuesta de hogares armonizada}
}
\value{
dataframe con variables seleccionadas
}
\description{
Verifica si el parametro x es una encuesta de hogares armonizada por SCLDATA, y devuelve las variables de vivienda
}
