% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimar_representacion_deficit.R
\name{estimar_representacion_deficit}
\alias{estimar_representacion_deficit}
\title{Estimacion de hogares segun estado habitacional}
\usage{
estimar_representacion_deficit(data, escala = "pais")
}
\arguments{
\item{data}{dataset de encuesta de hogares armonizada}

\item{escala}{nivel de analisis: "pais" o "region" .
La region esta determinada por la Encuesta de Hogar del pais de analisis}
}
\value{
dataset con la cantidad de hogares
}
\description{
Se estima la cantidad de hogares sin deficit, con deficit cuantitativo y con deficit cualitativo.
La estimacion se hace a partir del factor de expansion presente en las encuestas de hogares
}
